Added mouse functionality for rooms

+++ Mapping/Tools/MappingTool.cs
@@ -15,24 +15,24 @@ namespace Edelweiss.Mapping.Tools
         public virtual string DisplayName => Name.Substring(0, Name.Length - 4).CamelCaseToText();
 
         public virtual bool ClickingTriggersDrag => false;
-        internal void MouseDown(string room, int x, int y)
+        internal void MouseDown(string room, float x, float y)
         {
             if (ClickingTriggersDrag)
                 MouseDrag(room, x, y);
             MouseClick(room, x, y);
         }
 
-        public virtual void MouseClick(string room, int x, int y)
+        public virtual void MouseClick(string room, float x, float y)
         {
 
         }
 
-        public virtual void MouseDrag(string room, int x, int y)
+        public virtual void MouseDrag(string room, float x, float y)
         {
 
         }
 
-        public virtual void MouseRelease(string room, int x, int y)
+        public virtual void MouseRelease(string room, float x, float y)
         {
 
         }
